Move bing results processing utilities to their own file
